in_progress/streaming-test2/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
python-dotenv
google-generativeai
google-cloud-texttospeech
google-cloud-speech
opencv-python
numpy
# Add any other specific versions or packages if known e.g. google-adk
google-adk==1.0.0
certifi # Often needed for SSL with ADK
deprecated
